Refactor SQL Server keyword dictionaries and MsSQL leftovers
