+++ lib/Doctrine/DBAL/Driver/SQLSrv/Driver.php
@@ -50,7 +50,7 @@ class Driver implements \Doctrine\DBAL\Driver
 
     public function getDatabasePlatform()
     {
-        return new \Doctrine\DBAL\Platforms\SQLServer2008Platform();
+        return new \Doctrine\DBAL\Platforms\SQLServer2012Platform();
     }
 
     public function getSchemaManager(\Doctrine\DBAL\Connection $conn)
