+++ lib/Doctrine/DBAL/Tools/Console/Command/ReservedWordsCommand.php
@@ -31,7 +31,7 @@ class ReservedWordsCommand extends Command
 {
     private $keywordListClasses = array(
         'mysql'     => 'Doctrine\DBAL\Platforms\Keywords\MySQLKeywords',
-        'mssql'     => 'Doctrine\DBAL\Platforms\Keywords\MsSQLKeywords',
+        'sqlserver' => 'Doctrine\DBAL\Platforms\Keywords\SQLServerKeywords',
         'sqlite'    => 'Doctrine\DBAL\Platforms\Keywords\SQLiteKeywords',
         'pgsql'     => 'Doctrine\DBAL\Platforms\Keywords\PostgreSQLKeywords',
         'oracle'    => 'Doctrine\DBAL\Platforms\Keywords\OracleKeywords',
@@ -66,7 +66,7 @@ class ReservedWordsCommand extends Command
 Checks if the current database contains tables and columns
 with names that are identifiers in this dialect or in other SQL dialects.
 
-By default SQLite, MySQL, PostgreSQL, MsSQL and Oracle
+By default SQLite, MySQL, PostgreSQL, Microsoft SQL Server and Oracle
 keywords are checked:
 
     <info>%command.full_name%</info>
@@ -82,7 +82,7 @@ The following keyword lists are currently shipped with Doctrine:
     * pgsql
     * sqlite
     * oracle
-    * mssql
+    * sqlserver
     * db2 (Not checked by default)
 EOT
         );
@@ -98,7 +98,7 @@ EOT
 
         $keywordLists = (array)$input->getOption('list');
         if ( ! $keywordLists) {
-            $keywordLists = array('mysql', 'pgsql', 'sqlite', 'oracle', 'mssql');
+            $keywordLists = array('mysql', 'pgsql', 'sqlite', 'oracle', 'sqlserver');
         }
 
         $keywords = array();
